Moved extension code into its own sub directory
