+++ core/extension/Cargo.toml
@@ -1,0 +1,24 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+# SPDX-License-Identifier: MIT
+
+[package]
+name = "core-extension"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+
+[dependencies]
+core-protobuf-data-access = { path = "../protobuf_data_access" }
+log = { workspace = true }
+parking_lot = { workspace = true }
+strum = { workspace = true }
+strum_macros = { workspace = true }
+tokio = { workspace = true , features = ["macros", "rt-multi-thread"] }
+tonic = { workspace = true }
+
+[build-dependencies]
+tonic-build = { workspace = true }
+
+[features]
+managed_subscribe = []
